Enable sync_publish option for Qpid to combat timeouts.
The Qpid client was throwing SessionExceptions complaining that it
"timed out waiting for completion".  By adding the
sync_publisher='persistent' option to the connection URL, we can force
Qpid to sync after ever persistent object is sent.  The end result is
that the client blocks until the bus confirms receipt of the message.
Since the session is maintained due to the blocking, we don't have the
issue of timing out waiting to establish a session.

See
https://qpid.apache.org/releases/qpid-0.14/books/Programming-In-Apache-Qpid/html/ch03s02.html